+++ src/main/java/com/unesc/earthBnb/service/ComodidadeService.java
@@ -1,7 +1,7 @@
 package com.unesc.earthBnb.service;
 
-import com.unesc.earthBnb.dto.ComodidadeRequest;
-import com.unesc.earthBnb.dto.ComodidadeResponse;
+import com.unesc.earthBnb.dto.request.ComodidadeRequest;
+import com.unesc.earthBnb.dto.response.ComodidadeResponse;
 import com.unesc.earthBnb.exception.ComodidadeNaoEncontradoException;
 import com.unesc.earthBnb.mapper.ComodidadeMapper;
 import com.unesc.earthBnb.model.Comodidades;
@@ -51,6 +51,9 @@ public class ComodidadeService {
 
     @Transactional
     public void deleteComodidade(Long id) {
+        if (!comodidadeRepository.existsById(id)) {
+            throw new ComodidadeNaoEncontradoException(id);
+        }
         comodidadeRepository.deleteById(id);
     }
